Respond with 503 on full login request store
Addresses https://github.com/digital-asset/daml/pull/8351#pullrequestreview-560859604

+++ triggers/service/auth/src/test/scala/com/daml/auth/middleware/oauth2/TestFixture.scala
@@ -44,8 +44,8 @@ trait TestFixture
   self: Suite =>
   protected val ledgerId: String = "test-ledger"
   protected val jwtSecret: String = "secret"
-  protected val maxMiddlewareLogins: Long = Config.DefaultMaxLoginRequests
-  protected val maxClientAuthCallbacks: Long = 1000
+  protected val maxMiddlewareLogins: Int = Config.DefaultMaxLoginRequests
+  protected val maxClientAuthCallbacks: Int = 1000
   protected val middlewareCallbackUri: Option[Uri] = None
   lazy protected val clock: AdjustableClock = suiteResource.value.clock
   lazy protected val server: OAuthServer = suiteResource.value.authServer
